requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
praw==7.7.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
openai==1.12.0
requests==2.31.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
